Storage: warnings - careful with mixing 'unsigned int' with 'size_t' for offset/sizes, it invites a lot of warnings from 64-bit MSC, which can only be solved by ugly casts.
svn:sync-xref-src-repo-rev: r109310

+++ src/VBox/Storage/testcase/VDScriptAst.cpp
@@ -1,4 +1,4 @@
-/** $Id: VDScriptAst.cpp 109038 2016-07-22 18:30:37Z knut.osmundsen@oracle.com $ */
+/** $Id: VDScriptAst.cpp 109310 2016-07-30 11:48:53Z knut.osmundsen@oracle.com $ */
 /** @file
  *
  * VBox HDD container test utility - scripting engine AST node related functions.
@@ -346,7 +346,7 @@ DECLHIDDEN(PVDSCRIPTASTCORE) vdScriptAstNodeAlloc(VDSCRIPTASTCLASS enmClass)
     return pAstNode;
 }
 
-DECLHIDDEN(PVDSCRIPTASTIDE) vdScriptAstNodeIdeAlloc(unsigned cchIde)
+DECLHIDDEN(PVDSCRIPTASTIDE) vdScriptAstNodeIdeAlloc(size_t cchIde)
 {
     PVDSCRIPTASTIDE pAstNode = (PVDSCRIPTASTIDE)RTMemAllocZ(RT_OFFSETOF(VDSCRIPTASTIDE, aszIde[cchIde + 1]));
     if (pAstNode)
